Now with (veeeery rudimentary) animation!
Okay, here is a very simple example of what we might be able to expect
in terms of the animation?  Yeah, I know it still needs a lot of work.
Although my brother said that the graphics would not make him not buy
it, so that is a good start!