Screen stuff; look at extended desc. or Check-In
I've changed the structure somewhat to allow for simple and
straightforward screen switching in a manner that seems to be standard
practice.   DanForAllSeasons now extends Game and DemoScreen is one of a
few other screens in their own respective package.  Things still look a
little ghetto at the moment, but I'll try to clean it up when time
permits. I still need to work pausing along with a few other little
things (animating the splash screen instead of required a button prompt,
cleaning up the code, etc.)