Additional Art
Added a tree, bush, a non copyright-infringing beanstalk, the flower now
"dances" with two frames of animation.  Not sure if we actually wanted
to pursue that seriously or not.  Let me know what you think.  I suspect
the acorn might be too large, and I'm not happy with the "dirt" under
the hill right now, but it's good as a placeholder if nothing else.