Graphics Update
Okay, I fixed that pixel issue on the hills, made the grass thickness on
the hills better match that on the level terrain, and made tiles for
Fall and Winter as well.  I think the naming scheme is pretty
self-explanatory, but I'm not necessarily convinced it's the best, let
me know your thoughts.  We need to figure out how to substantially
differentiate Spring and Summer.  A possibility might be to make one of
them "pink" in theme.  Guess I'll try that.